add Company polymorphism

+++ 06_InheritanceAndPolymorphism/Company/src/Company/Employee.java
@@ -1,5 +1,6 @@
 package Company;
 
 public interface Employee {
-  double getMonthSalary();
+
+  Integer getMonthSalary();
 }
